Rename client package to dataprovider and re-structure
