--- tools/src/main/java/io/github/ygojson/tools/client/ClientConfig.java
@@ -1,26 +1,0 @@
-package io.github.ygojson.tools.client;
-
-import java.time.Duration;
-import java.util.function.Function;
-
-import io.github.ygojson.tools.common.ApplicationInfo;
-
-public interface ClientConfig<T> {
-	record RateLimit(int maxRequest, Duration duration) {}
-
-	String name();
-
-	Class<T> apiClass();
-
-	String baseUrl();
-
-	RateLimit rateLimit();
-
-	/**
-	 * Gets the supplier to build the user-agent given the current version of the tool.
-	 *
-	 * @param info information of the application.
-	 * @return supplier for the user-agent.
-	 */
-	Function<String, String> userAgentMapper(final ApplicationInfo info);
-}
