--- tools/src/main/java/io/github/ygojson/tools/client/yugipedia/params/SortDirection.java
@@ -1,19 +1,0 @@
-package io.github.ygojson.tools.client.yugipedia.params;
-
-import lombok.RequiredArgsConstructor;
-
-/**
- * Contains only sort directions used on YGOJSON.
- */
-@RequiredArgsConstructor
-public enum SortDirection {
-	OLDER("older"),
-	NEWER("newer");
-
-	private final String value;
-
-	@Override
-	public String toString() {
-		return value;
-	}
-}
