--- tools/src/main/java/io/github/ygojson/tools/client/yugipedia/response/Categories.java
@@ -1,10 +1,0 @@
-package io.github.ygojson.tools.client.yugipedia.response;
-
-import lombok.Data;
-
-@Data
-public class Categories {
-
-	private int ns;
-	private String title;
-}
