--- tools/src/main/java/io/github/ygojson/tools/client/yugipedia/response/Page.java
@@ -1,15 +1,0 @@
-package io.github.ygojson.tools.client.yugipedia.response;
-
-import java.util.List;
-
-import lombok.Data;
-
-@Data
-public class Page {
-
-	private long pageid;
-	private int ns;
-	private String title;
-	private List<Revision> revisions;
-	private List<Categories> categories;
-}
