--- tools/src/main/java/io/github/ygojson/tools/client/yugipedia/response/Revision.java
@@ -1,14 +1,0 @@
-package io.github.ygojson.tools.client.yugipedia.response;
-
-import java.time.ZonedDateTime;
-
-import lombok.Data;
-
-@Data
-class Revision {
-
-	private ZonedDateTime timestamp;
-	private String contentformat;
-	private String contentmodel;
-	private String content;
-}
