+++ tools/src/main/java/io/github/ygojson/tools/dataprovider/impl/yugipedia/model/api/Categories.java
@@ -1,0 +1,10 @@
+package io.github.ygojson.tools.dataprovider.impl.yugipedia.model.api;
+
+import lombok.Data;
+
+@Data
+public class Categories {
+
+	private int ns;
+	private String title;
+}
