+++ tools/src/main/java/io/github/ygojson/tools/dataprovider/impl/yugipedia/params/Category.java
@@ -1,0 +1,18 @@
+package io.github.ygojson.tools.dataprovider.impl.yugipedia.params;
+
+import lombok.RequiredArgsConstructor;
+
+/**
+ * Contains only categories used within YGOJSON.
+ */
+@RequiredArgsConstructor
+public enum Category {
+	CARDS("Category:Duel_Monsters_cards");
+
+	private final String value;
+
+	@Override
+	public String toString() {
+		return value;
+	}
+}
